requirements.txt
numpy~=1.19.4
torch~=1.9.1+cu111
trimesh~=3.9.29
tqdm~=4.45.0
argparse~=1.4.0
yaml~=0.1.7
pyyaml~=5.3.1
scipy~=1.4.1
pyrender~=0.1.43
pillow~=7.2.0
typing~=3.6.4
wandb~=0.12.2
tabulate~=0.8.7
pytorch-lightning~=1.4.9